// src/Optimism/utils/types.ts
export type DexType = 
  | "uniswapv3" 
  | "velodromeslipstream"
  | "solidlyv3"
  | "velodromefinancev2";
// ...
export interface DexPair {
  pairaddress?: string;
  dex: DexType;
  url?: string;
  tokenIn: {
    address: string;
    symbol: string;
    decimals:number;
  };
  tokenOut: {
    address: string;
    symbol: string;
    decimals: number;
  };
  liquidity?: {
    usd?: Number;
  };
  volume24h?: {
    usd?: number;
  };
  priceNative?: string;
  priceUsd?: string;
}